Replace deprecated (double) cast with (float) (#6631)
* fix: replace deprecated (double) cast with (float)

PHP 8.4 deprecates the non-canonical (double) cast. Replace all 11
instances with the canonical (float) equivalent. No behavioral change.


* fix: use float literals instead of no-op left-side casts

Left-side casts like (float) $var = 0 are no-ops — the cast is
discarded before assignment. Use 0.0 float literals instead.


---------

+++ lib/utility.php
@@ -1684,7 +1684,7 @@ function utilities_php_modules() : string {
 function memory_bytes(mixed $val) : mixed {
 	$val  = trim($val);
 	$last = strtolower($val[strlen($val) - 1]);
-	$val  = (double) trim($val, 'GMKgmk');
+	$val  = (float) trim($val, 'GMKgmk');
 
 	switch($last) {
 		case 'g':
